refactor: move ItemInfo files to new directory
