+++ mission/src/components/ItemInfo/Detail.vue
